+++ src/aws/ec2-wrapper.ts
@@ -1,11 +1,29 @@
-import { EC2 } from "aws-sdk"
-import { getAWSPagedResults } from "../utils"
+import { EC2 } from "aws-sdk";
+import { getAWSPagedResults } from "../utils";
+import { VPC, VPCDiscovery } from "../types";
 
-class EC2Wrapper {
+export class EC2Wrapper {
   public ec2: EC2
 
   constructor (credentials: any) {
-    this.ec2 = new EC2(credentials)
+    this.ec2 = new EC2(credentials);
+  }
+
+  /**
+   * Gets the desired vpc with the designated subnets and security groups
+   * that were set in serverless config file
+   * @returns {Promise<object>}
+   */
+  public async getVpcConfig (vpcDiscovery: VPCDiscovery): Promise<VPC> {
+    const vpc: VPC = {};
+    const vpcId = await this.getVpcId(vpcDiscovery.vpcName);
+    if (vpcDiscovery.subnetNames) {
+      vpc.subnetIds = await this.getSubnetIds(vpcId, vpcDiscovery.subnetNames);
+    }
+    if (vpcDiscovery.securityGroupNames) {
+      vpc.securityGroupIds = await this.getSecurityGroupIds(vpcId, vpcDiscovery.securityGroupNames);
+    }
+    return vpc;
   }
 
   /**
@@ -19,19 +37,19 @@ class EC2Wrapper {
         Name: "tag:Name",
         Values: [vpcName]
       }]
-    }
-    const vpcs = await getAWSPagedResults(
+    };
+    const vpcItems = await getAWSPagedResults(
       this.ec2,
       "describeVpcs",
       "Vpcs",
       "NextToken",
       "NextToken",
       params
-    )
-    if (vpcs.length === 0) {
-      throw new Error("Invalid vpc name, it does not exist")
+    );
+    if (vpcItems.length === 0) {
+      throw new Error("Invalid vpc name, it does not exist");
     }
-    return vpcs[0].VpcId
+    return vpcItems[0].VpcId;
   }
 
   /**
@@ -50,7 +68,7 @@ class EC2Wrapper {
         Name: "tag:Name",
         Values: subnetNames
       }]
-    }
+    };
 
     const subnets = await getAWSPagedResults(
       this.ec2,
@@ -59,22 +77,29 @@ class EC2Wrapper {
       "NextToken",
       "NextToken",
       params
-    )
+    );
 
     if (subnets.length === 0) {
-      throw new Error("Invalid subnet name, it does not exist")
+      throw new Error("Invalid subnet name, it does not exist");
     }
 
-    const missingSubnets = subnets.filter((subnet) => {
-      const nameTag = subnet.Tags.find(tag => tag.Key === "Name")
-      return subnetNames.indexOf(nameTag.Value) === -1
-    })
+    const missingSubnetNames = subnetNames.filter((subnetName) => {
+      // collect subnets by name
+      const subnetsByName = subnets.filter((subnet) => {
+        const nameTag = subnet.Tags.find((tag) => tag.Key === "Name");
+        return nameTag.Value === subnetName;
+      });
+      return subnetsByName.length === 0;
+    });
 
-    if (missingSubnets.length) {
-      throw new Error(`Not all subnets were registered: ${missingSubnets}`)
+    if (missingSubnetNames.length) {
+      throw new Error(
+        `Subnets do not exist for the names: ${missingSubnetNames}. ` +
+        "Please check the names are correct or remove it."
+      );
     }
 
-    return subnets.map(subnet => subnet.SubnetId)
+    return subnets.map((subnet) => subnet.SubnetId);
   }
 
   /**
@@ -92,7 +117,7 @@ class EC2Wrapper {
         Name: "group-name",
         Values: securityGroupNames
       }]
-    }
+    };
     const securityGroups = await getAWSPagedResults(
       this.ec2,
       "describeSecurityGroups",
@@ -100,22 +125,27 @@ class EC2Wrapper {
       "NextToken",
       "NextToken",
       params
-    )
+    );
 
     if (securityGroups.length === 0) {
-      throw new Error("Invalid security group name, it does not exist")
+      throw new Error("Invalid security group name, it does not exist");
     }
 
-    const missingGroups = securityGroups.filter((group) => {
-      return securityGroupNames.indexOf(group.GroupName) === -1
-    })
+    const missingGroupsNames = securityGroupNames.filter((groupName) => {
+      // collect security groups by name
+      const securityGroupsByName = securityGroups.filter((securityGroup) => {
+        return securityGroup.GroupName === groupName;
+      });
+      return securityGroupsByName.length === 0;
+    });
 
-    if (missingGroups.length) {
-      throw new Error(`Not all security group were registered: ${missingGroups}`)
+    if (missingGroupsNames.length) {
+      throw new Error(
+        `Security groups do not exist for the names: ${missingGroupsNames}. ` +
+        "Please check the names are correct or remove it."
+      );
     }
 
-    return securityGroups.map(group => group.GroupId)
+    return securityGroups.map((group) => group.GroupId);
   }
 }
-
-export = EC2Wrapper