+++ src/index.ts
@@ -1,20 +1,23 @@
-"use strict"
+"use strict";
 
-import { ServerlessInstance } from "./types"
-import EC2Wrapper = require("./aws/ec2-wrapper")
+import { ServerlessInstance } from "./types";
+import { LambdaFunction } from "./common/lambda-function";
+import Globals from "./globals";
 
 class VPCPlugin {
   private serverless: ServerlessInstance;
   public hooks: object;
-  public ec2Wrapper: EC2Wrapper;
+  public awsCredentials: any;
+  public lambdaFunction: LambdaFunction;
 
   constructor (serverless) {
-    this.serverless = serverless
+    this.serverless = serverless;
+    Globals.serverless = serverless;
 
     /* hooks are the actual code that will run when called */
     this.hooks = {
-      "before:package:initialize": this.hookWrapper.bind(this, this.updateVpcConfig)
-    }
+      "before:package:initialize": this.hookWrapper.bind(this, this.updateFunctionsVpcConfig)
+    };
   }
 
   /**
@@ -22,52 +25,94 @@ class VPCPlugin {
    * @param lifecycleFunc lifecycle function that actually does desired action
    */
   public async hookWrapper (lifecycleFunc: any) {
-    this.validateConfigExists()
-    this.initAWSResources()
-    return await lifecycleFunc.call(this)
+    this.validateConfigExists();
+    this.initResources();
+    return await lifecycleFunc.call(this);
   }
 
   /**
    * Validate if the plugin config exists
    */
   public validateConfigExists (): void {
-    const config = this.serverless.service.custom
-    const vpc = config && config.vpc
+    const config = this.serverless.service.custom;
+    if (config && !config.vpcDiscovery) {
+      config.vpcDiscovery = config.vpc;
+      Globals.logWarning(
+        "The `vpc` option of `custom` config is deprecated and will be removed in the future. " +
+        "Please use `vpcDiscovery` option instead."
+      );
+    }
+    const vpc = config && config.vpcDiscovery;
     // Checks if the serverless file is setup correctly
-    if (!vpc || vpc.vpcName == null || vpc.subnetNames == null || vpc.securityGroupNames == null) {
-      throw new Error("Serverless file is not configured correctly. Please see README for proper setup.")
+    if (!vpc || vpc.vpcName == null || (vpc.subnetNames == null && vpc.securityGroupNames == null)) {
+      throw new Error(
+        "Serverless file is not configured correctly. " +
+        "You must specify the vpcName and at least one of subnetNames or securityGroupNames. " +
+        "Please see README for proper setup."
+      );
     }
   }
 
   /**
    * Setup AWS resources
    */
-  public initAWSResources (): void {
-    const credentials = this.serverless.providers.aws.getCredentials()
-    credentials.region = this.serverless.providers.aws.getRegion()
-    this.ec2Wrapper = new EC2Wrapper(credentials)
+  public initResources (): void {
+    this.awsCredentials = this.serverless.providers.aws.getCredentials();
+    this.awsCredentials.region = this.serverless.providers.aws.getRegion();
+    this.lambdaFunction = new LambdaFunction(this.awsCredentials, this.serverless.service.custom.vpcDiscovery);
   }
 
   /**
-   * Gets the desired vpc with the designated subnets and security groups
-   * that were set in serverless config file
+   * Updates functions vpc config
    * @returns {Promise<object>}
    */
-  public async updateVpcConfig (): Promise<object> {
-    this.serverless.cli.log("Updating VPC config...")
+  public async updateFunctionsVpcConfig (): Promise<object> {
+    Globals.logInfo("Updating VPC config...");
+    const service = this.serverless.service;
 
-    const service = this.serverless.service
+    // Sets the serverless's vpc config
+    if (service.functions) {
+      // loop through the functions and update VPC config
+      // eslint-disable-next-line guard-for-in
+      for (const funcName in service.functions) {
+        // eslint-disable-next-line
+        const func = service.functions[funcName];
+        const funcVPC = await this.lambdaFunction.getFuncVPC(funcName, func.vpcDiscovery);
 
-    const vpcId = await this.ec2Wrapper.getVpcId(service.custom.vpc.vpcName)
-    const subnetIds = await this.ec2Wrapper.getSubnetIds(vpcId, service.custom.vpc.subnetNames)
-    const securityGroupIds = await this.ec2Wrapper.getSecurityGroupIds(vpcId, service.custom.vpc.securityGroupNames)
+        if (!funcVPC) {
+          continue;
+        }
 
-    service.provider.vpc = {
-      subnetIds: subnetIds,
-      securityGroupIds: securityGroupIds
+        // init vpc empty config in case not exists
+        func.vpc = func.vpc || {};
+        // log warning in case vpc.subnetIds and vpcDiscovery.subnetNames are specified.
+        if (func.vpc.subnetIds && func.vpcDiscovery && func.vpcDiscovery.subnetNames) {
+          Globals.logWarning(
+            `vpc.subnetIds' are specified for the function '${funcName}' 
+            and overrides 'vpcDiscovery.subnetNames' discovery config.`
+          );
+        }
+        // log warning in case vpc.securityGroupIds and vpcDiscovery.securityGroupNames are specified.
+        if (func.vpc.securityGroupIds && func.vpcDiscovery && func.vpcDiscovery.securityGroupNames) {
+          Globals.logWarning(
+            `vpc.securityGroupIds' are specified for the function '${funcName}' 
+            and overrides 'vpcDiscovery.securityGroupNames' discovery config.`
+          );
+        }
+
+        // set vpc.subnetIds if it does not exists
+        if (!func.vpc.subnetIds && funcVPC.subnetIds) {
+          func.vpc.subnetIds = funcVPC.subnetIds;
+        }
+
+        // set vpc.securityGroupIds if it does not exists
+        if (!func.vpc.securityGroupIds && funcVPC.securityGroupIds) {
+          func.vpc.securityGroupIds = funcVPC.securityGroupIds;
+        }
+      }
     }
 
-    return service.provider.vpc
+    return service.functions;
   }
 }