+++ test/integration-tests/utils/execution.ts
@@ -1,6 +1,6 @@
-"use strict"
+"use strict";
 
-import shell from "shelljs"
+import shell from "shelljs";
 
 /**
  * Executes given shell command.
@@ -8,16 +8,16 @@ import shell from "shelljs"
  * @returns {Promise<void>} Resolves if successfully executed, else rejects
  */
 async function exec (cmd) {
-  console.debug(`\tRunning command: ${cmd}`)
+  console.debug(`\tRunning command: ${cmd}`);
   return new Promise((resolve, reject) => {
     shell.exec(cmd, { silent: false }, (err, stdout, stderr) => {
       if (err || stderr) {
         // eslint-disable-next-line prefer-promise-reject-errors
-        return reject()
+        return reject();
       }
-      return resolve()
-    })
-  })
+      return resolve();
+    });
+  });
 }
 
 /**
@@ -26,14 +26,14 @@ async function exec (cmd) {
  * @param {string} folderName
  */
 async function createTempDir (tempDir, folderName) {
-  await exec(`rm -rf ${tempDir}`)
-  await exec(`mkdir -p ${tempDir} && cp -R test/integration-tests/${folderName}/. ${tempDir}`)
-  await exec(`mkdir -p ${tempDir}/node_modules/.bin`)
-  await exec(`ln -s $(pwd) ${tempDir}/node_modules/`)
+  await exec(`rm -rf ${tempDir}`);
+  await exec(`mkdir -p ${tempDir} && cp -R test/integration-tests/${folderName}/. ${tempDir}`);
+  await exec(`mkdir -p ${tempDir}/node_modules/.bin`);
+  await exec(`ln -s $(pwd) ${tempDir}/node_modules/`);
 
-  await exec(`ln -s $(pwd)/node_modules/serverless ${tempDir}/node_modules/`)
+  await exec(`ln -s $(pwd)/node_modules/serverless ${tempDir}/node_modules/`);
   // link serverless to the bin directory so we can use $(npm bin) to get the path to serverless
-  await exec(`ln -s $(pwd)/node_modules/serverless/bin/serverless.js ${tempDir}/node_modules/.bin/serverless`)
+  await exec(`ln -s $(pwd)/node_modules/serverless/bin/serverless.js ${tempDir}/node_modules/.bin/serverless`);
 }
 
 /**
@@ -43,7 +43,7 @@ async function createTempDir (tempDir, folderName) {
  * @returns {Promise<void>}
  */
 function slsDeploy (tempDir, identifier) {
-  return exec(`cd ${tempDir} && $(npm bin)/serverless deploy --RANDOM_STRING ${identifier}`)
+  return exec(`cd ${tempDir} && $(npm bin)/serverless deploy --RANDOM_STRING ${identifier}`);
 }
 
 /**
@@ -53,7 +53,7 @@ function slsDeploy (tempDir, identifier) {
  * @returns {Promise<void>}
  */
 function slsRemove (tempDir, identifier) {
-  return exec(`cd ${tempDir} && $(npm bin)/serverless remove --RANDOM_STRING ${identifier}`)
+  return exec(`cd ${tempDir} && $(npm bin)/serverless remove --RANDOM_STRING ${identifier}`);
 }
 
 export {
@@ -61,4 +61,4 @@ export {
   exec,
   slsDeploy,
   slsRemove
-}
+};